requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pygithub==2.1.1
gitpython==3.1.40
cryptography==41.0.7
python-dotenv==1.0.0
requests==2.31.0 